requirements.txt
 Django==2.1.5
django-pure-pagination==0.3.0
Pillow==6.2.1
PyMySQL==0.9.3

